add new useRelationship hook
